gateway/requirements.txt
alembic>=1.10.2
autopep8>=2.0.2
cffi>=1.15.1
click>=8.0.3
cryptography>=41.0.3
Flask>=2.2.5
Flask-Migrate>=4.0.4
Flask-SQLAlchemy>=3.0.3
greenlet>=2.0.2
itsdangerous>=2.0.1
Jinja2>=3.0.2
Mako>=1.2.4
MarkupSafe>=2.1.2
paho-mqtt>=1.6.1
protobuf>=4.22.1
psycopg2-binary>=2.9.5
pybgapi>=1.2.0
pycodestyle>=2.10.0
pycparser>=2.21
pyOpenSSL>=23.2.0
pyserial>=3.5
SQLAlchemy>=2.0.5.post1
tomli>=2.0.1
typing_extensions>=4.5.0
Werkzeug>=2.2.3

